requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dj-rest-auth==7.0.1
django-allauth==65.13.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
requests==2.32.5
drf-yasg==1.21.11
whitenoise==6.11.0
gunicorn==23.0.0
Pillow==12.0.0
psycopg-binary==3.3.2
psycopg[binary]>=3.1.8